Make setup easier to use, and make tabster not required
