+++ ember-primitives/package.json
@@ -104,9 +104,7 @@
     "version": 2,
     "type": "addon",
     "main": "addon-main.cjs",
-    "app-js": {
-      "./services/ember-primitives/setup.js": "./dist/_app_/services/ember-primitives/setup.js"
-    }
+    "app-js": {}
   },
   "exports": {
     ".": {
