+++ ember-primitives/src/-private.ts
@@ -1,0 +1,4 @@
+/**
+ * @internal
+ */
+export const PRIMITIVES = Symbol.for('ember-primitives-globals');
